split auth service
