+++ data-reconciliation-app/Makefile
@@ -92,4 +92,5 @@ generate-jwt-token: # Generate an Azure AD access tokens for authentication
 	@echo "########################################################################"
 	@curl -X POST https://login.microsoftonline.com/16b3c013-d300-468d-ac64-7eda0820b6d3/oauth2/v2.0/token \
 	-H "Content-Type: application/x-www-form-urlencoded" \
-	-d "client_id=${Client_Id}&client_secret=${Client_Secret}&scope=api://b8dbd573-a015-424b-b111-2d5fa11cee3c/.default&grant_type=client_credentials"
+	-d "client_id=${Client_Id}&client_secret=${Client_Secret}&scope=api://b8dbd573-a015-424b-b111-2d5fa11cee3c/.default
+	&grant_type=client_credentials" | jq -r .access_token
